Add new files

+++ lec_3_Conditional_Statements/06_Number_100_To_200.py
@@ -1,4 +1,4 @@
-number = int(input())
+number = float(input())
 
 if number < 100:
     print("Less than 100")
